app/requirements.txt
fastapi
pydantic
PyJWT
uvicorn
python-dotenv
